Add files via upload
In this project I have cleaned the data. In this data there were unwanted data like review_title, about product, image links, product_link, and review_content were dropped from the data. Then checked the data types of each column and found that the data types of each column and changed the data types of column which are wrong. Splitted the Category into three naming Category 1, Category 2, Category 3. Visualized the data according to categories which have highest discount price and so on.